fix: harden — 35 production gaps across 3 adversarial rounds
Round 1 (15 fixes):
- Weight sync: partial sync validation, vLLM traversal check, always synchronize
- Advantages: batch_contexts validation, step_region_map guard, virtual step init
- Generation: logprobs fallback -inf, padding -inf, BF16→FP32 before log_softmax
- Training: NaN check before backward, per-token metrics tracking, masked_mean types
- Config: int() validation with helpful errors, reward_fn .get() access

Round 2 (12 fixes):
- Weight sync: multi-GPU synchronize device arg
- Advantages: aspiration warmup getattr, EMA variance new_mean fix (critical),
  batch_contexts None check, critic NaN .item() guard, requires_grad warning
- Generation: Triton bounds check, sum_exp underflow clamp
- Training: accumulated loss checkpoint save, _resumed_mid_accumulation flag
- Config: empty train_files validation, FileNotFoundError context

Round 3 (8 fixes):
- Advantages: variance state merge on load, NaN filter before stack
- Generation: Triton empty seq_len guard, max_logprobs patch error context
- Training: -inf→-1e9 padding to avoid inf KL
- Config: list type validation for metadata_columns, initial_tiers, tier_order

Leverage (Opus): 51% of findings trace to missing type-safe data flow at
module boundaries. Proposed: typed DTOs, frozen dataclasses, tensor helpers.

+++ qgre/nemo_extracted/logits.py
@@ -48,14 +48,14 @@ def selective_log_softmax(
         selected = torch.gather(logits, dim=-1, index=index.unsqueeze(-1)).squeeze(-1)
         return (selected - lse).to(torch.float32)
     else:
-        # bf16/fp16: logsumexp loses precision. Use per-row log_softmax instead.
-        # Still memory-efficient: processes one row at a time.
+        # GEN-R1-5: bf16/fp16: convert to FP32 BEFORE log_softmax for stability
+        # logsumexp in BF16 loses precision — compute log_softmax in FP32.
         token_logprobs = torch.zeros_like(index, dtype=torch.float32)
         for i, (logits_row, index_row) in enumerate(zip(logits, index)):
-            logprobs_row = logits_row.log_softmax(dim=-1)
+            logprobs_row = logits_row.float().log_softmax(dim=-1)
             token_logprobs[i] = torch.gather(
                 logprobs_row, dim=-1, index=index_row.unsqueeze(-1)
-            ).squeeze(-1).to(torch.float32)
+            ).squeeze(-1)
         return token_logprobs
